completed billings plus user permissions module

+++ database/migrations/2018_03_28_151844_create_visits_table.php
@@ -13,23 +13,25 @@ class CreateVisitsTable extends Migration
      */
     public function up()
     {
-        Schema::create('visits', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id')->index()->nullable();
-            $table->unsignedInteger('patient_id')->index();
-            $table->string('type');
-            $table->string('title');
-            $table->text('subjects')->nullable();
-            $table->text('objects')->nullable();
-            $table->text('assessment')->nullable();
-            $table->text('plans')->nullable();
-            $table->boolean('successful_delivery')->default(0);
-            $table->datetime('discharged_on')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('visits')) {
+            Schema::create('visits', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('user_id')->index()->nullable();
+                $table->unsignedInteger('patient_id')->index();
+                $table->string('type');
+                $table->string('title');
+                $table->text('subjects')->nullable();
+                $table->text('objects')->nullable();
+                $table->text('assessment')->nullable();
+                $table->text('plans')->nullable();
+                $table->boolean('successful_delivery')->default(0);
+                $table->datetime('discharged_on')->nullable();
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->foreign('patient_id')->references('id')->on('patients')->onDelete('cascade');
-        });
+                $table->foreign('user_id')->references('id')->on('users');
+                $table->foreign('patient_id')->references('id')->on('patients')->onDelete('cascade');
+            });
+        }
     }
 
     /**
