+++ database/migrations/2018_04_06_191542_create_surgeries_table.php
@@ -13,20 +13,22 @@ class CreateSurgeriesTable extends Migration
      */
     public function up()
     {
-        Schema::create('surgeries', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id')->index()->nullable();
-            $table->unsignedInteger('visit_id')->index();
-            $table->unsignedInteger('surgery_id')->index()->nullable();
-            $table->string('surgery_name');
-            $table->date('surgery_date')->nullable();
-            $table->text('complications')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('surgeries')) {
+            Schema::create('surgeries', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('user_id')->index()->nullable();
+                $table->unsignedInteger('visit_id')->index();
+                $table->unsignedInteger('surgery_id')->index()->nullable();
+                $table->string('surgery_name');
+                $table->date('surgery_date')->nullable();
+                $table->text('complications')->nullable();
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->foreign('visit_id')->references('id')->on('visits')->onDelete('cascade');
-            $table->foreign('surgery_id')->references('id')->on('surgeries')->onDelete('cascade');
-        });
+                $table->foreign('user_id')->references('id')->on('users');
+                $table->foreign('visit_id')->references('id')->on('visits')->onDelete('cascade');
+                $table->foreign('surgery_id')->references('id')->on('surgeries')->onDelete('cascade');
+            });
+        }
     }
 
     /**
